requirements.txt
fastapi
uvicorn[standard]
redis
motor
python-dotenv
pika
kafka-python
pydantic